+++ .travis.yml
@@ -2,10 +2,8 @@ language: android
 dist: trusty
 
 before_install:
-  - openssl aes-256-cbc -K $encrypted_a039e732823a_key -iv $encrypted_a039e732823a_iv -in travis/secrets.tar.enc -out travis/secrets.tar -d
-  - tar xvf travis/secrets.tar
-  - cp travis/google-services.json app/google-services.json
-  - cp travis/home-assistant-mobile-apps-5fd6b9dd0fdb.json serviceCredentialsFile.json
+  - chmod ugo+x travis/before_install.sh
+  - ./travis/before_install.sh
 
 android:
   components:
