+++ app/build.gradle
@@ -15,7 +15,7 @@ android {
     compileSdkVersion 29
 
     defaultConfig {
-        applicationId "io.homeassistant.android"
+        applicationId "io.homeassistant.companion.android"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode "${System.env.VERSION_CODE ?: 1}".toInteger()
@@ -25,6 +25,26 @@ android {
     firebaseAppDistribution {
         serviceCredentialsFile = "serviceCredentialsFile.json"
         releaseNotesFile = "CHANGES.md"
+    }
+
+    signingConfigs {
+        release {
+            storeFile file('../release_keystore.keystore')
+            storePassword "${System.env.KEYSTORE_PASSWORD ?: ""}"
+            keyAlias "$System.env.KEYSTORE_ALIAS" ?: ""
+            keyPassword "${System.env.KEYSTORE_ALIAS_PASSWORD ?: ""}"
+            v1SigningEnabled true
+            v2SigningEnabled true
+        }
+    }
+
+    buildTypes {
+        release {
+            debuggable false
+            jniDebuggable false
+            zipAlignEnabled true
+            signingConfig signingConfigs.release
+        }
     }
 }
