--- app/src/main/java/io/homeassistant/android/webview/WebView.kt
@@ -1,12 +1,0 @@
-package io.homeassistant.android.webview
-
-import io.homeassistant.android.io.homeassistant.android.api.Token
-
-
-interface WebView {
-
-    fun loadUrl(url: String)
-
-    fun setToken(callback: String, token: Token)
-
-}
