+++ app/src/main/java/io/homeassistant/companion/android/HomeAssistantApplication.kt
@@ -1,7 +1,7 @@
-package io.homeassistant.android
+package io.homeassistant.companion.android
 
 import android.app.Application
-import io.homeassistant.android.api.Session
+import io.homeassistant.companion.android.api.Session
 
 class HomeAssistantApplication : Application() {
