+++ app/src/main/java/io/homeassistant/companion/android/api/AuthorizationCode.kt
@@ -1,13 +1,15 @@
-package io.homeassistant.android.api
+package io.homeassistant.companion.android.api
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
 
-data class RefreshToken(
+data class AuthorizationCode(
     @JsonProperty("access_token")
     val accessToken: String,
     @JsonProperty("expires_in")
     val expiresIn: Int,
+    @JsonProperty("refresh_token")
+    val refreshToken: String,
     @JsonProperty("token_type")
     val tokenType: String
 )
