+++ app/src/main/java/io/homeassistant/companion/android/api/RefreshToken.kt
@@ -1,15 +1,13 @@
-package io.homeassistant.android.api
+package io.homeassistant.companion.android.api
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
 
-data class AuthorizationCode(
+data class RefreshToken(
     @JsonProperty("access_token")
     val accessToken: String,
     @JsonProperty("expires_in")
     val expiresIn: Int,
-    @JsonProperty("refresh_token")
-    val refreshToken: String,
     @JsonProperty("token_type")
     val tokenType: String
 )
