+++ app/src/main/java/io/homeassistant/companion/android/onboarding/ManualSetupFragment.kt
@@ -1,4 +1,4 @@
-package io.homeassistant.android.onboarding
+package io.homeassistant.companion.android.onboarding
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
 import androidx.fragment.app.Fragment
-import io.homeassistant.android.R
+import io.homeassistant.companion.android.R
 
 
 class ManualSetupFragment : Fragment() {
