+++ app/src/main/java/io/homeassistant/companion/android/onboarding/ManualSetupListener.kt
@@ -1,4 +1,4 @@
-package io.homeassistant.android.onboarding
+package io.homeassistant.companion.android.onboarding
 
 
 interface ManualSetupListener {
