+++ app/src/main/java/io/homeassistant/companion/android/onboarding/OnboardingActivity.kt
@@ -1,12 +1,13 @@
-package io.homeassistant.android.onboarding
+package io.homeassistant.companion.android.onboarding
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import io.homeassistant.android.onboarding.authentication.AuthenticationFragment
-import io.homeassistant.android.onboarding.authentication.AuthenticationListener
-import io.homeassistant.android.webview.WebViewActivity
+import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.onboarding.authentication.AuthenticationFragment
+import io.homeassistant.companion.android.onboarding.authentication.AuthenticationListener
+import io.homeassistant.companion.android.webview.WebViewActivity
 
 
 class OnboardingActivity : AppCompatActivity(), DiscoveryListener, ManualSetupListener, AuthenticationListener {
@@ -21,12 +22,12 @@ class OnboardingActivity : AppCompatActivity(), DiscoveryListener, ManualSetupLi
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(io.homeassistant.android.R.layout.activity_onboarding)
+        setContentView(R.layout.activity_onboarding)
 
         if (savedInstanceState == null) {
             supportFragmentManager
                 .beginTransaction()
-                .add(io.homeassistant.android.R.id.content, DiscoveryFragment.newInstance())
+                .add(R.id.content, DiscoveryFragment.newInstance())
                 .commit()
         }
     }
@@ -42,7 +43,7 @@ class OnboardingActivity : AppCompatActivity(), DiscoveryListener, ManualSetupLi
     override fun onSelectManualSetup() {
         supportFragmentManager
             .beginTransaction()
-            .replace(io.homeassistant.android.R.id.content, ManualSetupFragment.newInstance())
+            .replace(R.id.content, ManualSetupFragment.newInstance())
             .commit()
     }
 
@@ -53,7 +54,7 @@ class OnboardingActivity : AppCompatActivity(), DiscoveryListener, ManualSetupLi
     override fun onSelectUrl(url: String) {
         supportFragmentManager
             .beginTransaction()
-            .replace(io.homeassistant.android.R.id.content, AuthenticationFragment.newInstance(url))
+            .replace(R.id.content, AuthenticationFragment.newInstance(url))
             .commit()
     }
