+++ app/src/main/java/io/homeassistant/companion/android/settings/SettingsActivity.kt
@@ -1,12 +1,12 @@
-package io.homeassistant.android.settings
+package io.homeassistant.companion.android.settings
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import io.homeassistant.android.BuildConfig
-import io.homeassistant.android.R
+import io.homeassistant.companion.android.BuildConfig
+import io.homeassistant.companion.android.R
 
 
 class SettingsActivity : AppCompatActivity() {
