+++ app/src/main/java/io/homeassistant/companion/android/webview/WebView.kt
@@ -1,0 +1,12 @@
+package io.homeassistant.companion.android.webview
+
+import io.homeassistant.companion.android.api.Token
+
+
+interface WebView {
+
+    fun loadUrl(url: String)
+
+    fun setToken(callback: String, token: Token)
+
+}
