+++ app/src/main/java/io/homeassistant/companion/android/webview/WebViewPresenterImpl.kt
@@ -1,10 +1,10 @@
-package io.homeassistant.android.webview
+package io.homeassistant.companion.android.webview
 
 import android.util.Log
-import io.homeassistant.android.api.AuthenticationService
-import io.homeassistant.android.api.HomeAssistantApi
-import io.homeassistant.android.api.RefreshToken
-import io.homeassistant.android.api.Session
+import io.homeassistant.companion.android.api.AuthenticationService
+import io.homeassistant.companion.android.api.HomeAssistantApi
+import io.homeassistant.companion.android.api.RefreshToken
+import io.homeassistant.companion.android.api.Session
 import retrofit2.Response
 import java.io.IOException
